lmrob.fit {robustbase}	R Documentation

MM-type estimator for regression

Description

Compute MM-type estimators of regression: An S-estimator is used as starting value, and an M-estimator with fixed scale and redescending psi-function is used from there. Optionally a D-step (Design Adaptive Scale estimate) as well as a second M-step is calculated.

Usage

lmrob.fit(x, y, control, init = NULL, mf = NULL, bare.only = FALSE)

Arguments

x
design matrix (n \times p) typically including a column of 1s for the intercept.

y
numeric response vector (of length n).

control
a list of control parameters as returned by lmrob.control, used for both the initial S-estimate and the subsequent M- and D-estimates.

init
optional list of initial estimates. See Details.

mf
defunct.

bare.only
logical indicating if the result should be return()ed after the bare computation steps are done. Useful, e.g., when you only need the coefficients.

Details

This function is the basic fitting function for MM-type estimation, called by lmrob and typically not to be used on its own.

If given, init must be a list of initial estimates containing at least the initial coefficients and scale as coefficients and scale. Otherwise it calls lmrob.S(..) and uses it as initial estimator.

Value

A list with components (some missing in case bare.only is true)

fitted.values
X \beta, i.e., X %*% coefficients.

residuals
the raw residuals, y - fitted.values

rweights
robustness weights derived from the final M-estimator residuals (even when not converged).

rank

degree.freedom
n - rank

coefficients
estimated regression coefficient vector

scale
the robustly estimated error standard deviation

cov
variance-covariance matrix of coefficients, if the RWLS iterations have converged (and control$cov is not "none").

control

iter

converged
logical indicating if the RWLS iterations have converged.

init.S
the whole initial S-estimator result, including its own converged flag, see lmrob.S (only for MM-estimates).

init
A similar list that contains the results of intermediate estimates (not for MM-estimates).

Author(s)

Matias Salibian-Barrera, Martin Maechler and Manuel Koller

See Also

lmrob, lmrob..M..fit, lmrob..D..fit, lmrob.S

[Package robustbase version 0.99-7 Index]