lmrob.fit.MM {robustbase}	R Documentation

MM-estimator for regression

Description

Compute MM-estimators of regression: An S-estimator is used as starting value, and an M-estimator with fixed scale and redescending psi-function is used from there.

Usage

lmrob.fit.MM(x, y, control)

Arguments

x 	 design matrix (n x p) typically including a column of 1s for the intercept.

y 	 numeric response vector (of length n).

control 	 A list of control parameters as returned by lmrob.control, used for both the initial S-estimate and the subsequent M-estimate.

Details

This function is the basic fitting function for MM-estimation, called by lmrob and typically not to be used on its own.

It calls lmrob.S(..) and uses it as initial estimator. Note that the inference used (covariance matrix) depends crucially on the S-estimator used, and hence it is currently no longer possible to specify the S-estimator at this level.

Value

A list with components

fitted.values 	 X beta, i.e. X %*% coefficients.

residuals 	 the raw residuals, y - fitted.values

weights 	 robustness weights derived from the final M-estimator residuals (even when not converged).

rank

degree.freedom 	 n - rank

coefficients 	 estimated regression coefficient vector

initial.coefficients

scale 	 the robustly estimated error standard deviation

cov 	 variance-covariance matrix of coefficients, if the RWLS iterations have converged, otherwise the vcov-matrix of the initial estimator.

control

iter

converged 	 logical indicating if the RWLS iterations have converged.

init.S 	 the whole initial S-estimator result, including its own converged flag, see lmrob.S.

See Also

lmrob, lmrob..M..fit, lmrob.S

[Package robustbase version 0.4-5 Index]