los {robustbase}	R Documentation

Length of Stay Data

Description

Length of stay for 201 patients that stayed at the University Hospital of Lausanne during the year 2000.

Usage

data(los)

Format

Vector of integer values giving the length of stay (days):

int [1:201] 16 13 17 4 15 24 59 18 33 8 ...

Details

These data may be used to estimate and predict the total resource consumption of this group of patients.

Cf. Ruffieux, Paccaud and Marazzi (2000).

Source

The data were kindly provided by A. Marazzi.

Cf. Hubert, M. and Vandervieren, E. (2006), p. 13–15.

References

Ruffieux, C., Paccaud, F. and A. Marazzi (2000) Comparing rules for truncating hospital length of stay; Casemix Quarterly 2, n. 1.

Hubert, M. and Vandervieren, E. (2006) An Adjusted Boxplot for Skewed Distributions, Technical Report TR-06-11, KU Leuven, Section of Statistics, Leuven.
http://wis.kuleuven.be/stat/robust/Papers/TR0611.pdf

Examples

 summary(los) # quite skewed, with median(.) = 8
 plot(table(los))
 boxplot(los, horizontal=TRUE, add=TRUE, col = "red", axes=FALSE)
 ##-> "outliers" instead of "just skewed"

 hist(log(los))
 boxplot(log(los), add=TRUE, col=2, border=2, horizontal = TRUE, at = -1)

 if(R.version$arch == "x86_64") { ## FIXME -- bug in C code

   ## Hubert and Vandervieren (2006), p. 15, Fig. 11.
   adjbox(los, col = "gray", outcol = "red")
   boxplot(los, add = TRUE, staplewex = 0.2, outcex = 0.5, outpch = 4)
   title("adjbox(los) and boxplot(los) superimposed")
 }

[Package robustbase version 0.4-5 Index]