plot-methods {robustbase}	R Documentation

Plot an Object of the "Psi Function" Class

Description

The plot method objects of class psi_func simply visualizes the ρ(), ψ(), and weight functions and their derivatives.

Usage

## S4 method for signature 'psi_func'
plot(x, y,
     which = c("rho", "psi", "Dpsi", "wgt", "Dwgt"),
     main = "full", 
     col = c("black", "red3", "blue3", "dark green", "light green"),
     leg.loc = "right", ...)

Arguments

x
object of class psi_func to be plotted

y
(optional) vector of abscissa values (to plot object at).

which
character vector of slots to be included in plot; by default, all of the slots are included

main
string or logical indicating the kind of plot title; either "full", "short" or FALSE which chooses a full, a short or no main title at all.

col
colors to be used for the different slots

leg.loc
legend placement, see also x argument of legend

...
passed to matplot

Note

An earlier version had argument shortMain which is deprecated now. Use main = "short" instead of shortMain = TRUE.

If you want to specify your own title, use main=FALSE, and a subsequent title(...) call.

See Also

psiFunc() and the class psi_func.

Examples

plot(huberPsi)
plot(huberPsi, which=c("psi", "Dpsi", "wgt"),
     main="short", leg = "topleft")

plot(hampelPsi)
## Plotting aspect ratio = 1:1 :
plot(hampelPsi, asp=1, main="short",
     which = c("psi", "Dpsi", "wgt", "Dwgt"))

[Package robustbase version 0.95-0 Index]