print.lmrob {robustbase}	R Documentation

Print Method for Objects of Class "lmrob"

Description

Print method for elements of class "lmrob".

Usage

## S3 method for class 'lmrob'
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

x
an R object of class lmrob, typically created by lmrob.

digits
number of digits for printing, see digits in options.

...
potentially more arguments passed to methods.

See Also

lmrob, summary.lmrob, print and summary.

Examples

data(coleman)
( m1 <- lmrob(Y ~ ., data=coleman) ) # -> print.lmrob() method

[Package robustbase version 0.95-0 Index]