psi_func-class {robustbase}	R Documentation

Class of "Psi Functions" for M-Estimation

Description

The class "psi_func" is used to store \psi \ (psi) functions for M-estimation. In particular, an object of the class contains \rho(x) \ (\code{rho}), its derivative \psi(x) \ (psi), the weight function \psi(x)/x, and first derivative of \psi, Dpsi = \psi'(x).

Objects from the Class

Objects can be created by calls of the form new("psi_func", ...), but preferably by psiFunc(...).

Slots

rho:

the \rho() function, an object of class "functionX". This is used to formulate the objective function; \rho() can be regarded as generalized negative log-likelihood.

psi:

\psi() is the derivative of \rho, \psi(x) = \frac{d}{dx} \rho(x); also of class "functionX".

wgt:

The weight function \psi(x)/x, of class "functionX".

Dpsi:

the derivative of \psi, Dpsi(x) = psi'(x); of class "functionX".

Dwgt:

the derivative of the weight function, of class "functionX", is generated automatically if psiFunc constructor is used.

tDefs:

named numeric vector of tuning parameter Default values.

Erho:

A function of class "functionXal" for computing E[\rho(X)] when X is standard normal \mathcal{N}(0,1).

Epsi2:

A function of class "functionXal" for computing E[\psi^2(X)] when X is standard normal.

EDpsi:

A function of class "functionXal" for computing E[\psi'(X)] when X is standard normal.

name:

Name of \psi-function used for printing.

xtras:

Potentially further information.

Methods

Currently, only chgDefaults(), plot() and show().

Author(s)

Martin Maechler

See Also

psiFunc.

Examples

str(huberPsi, give.attr = FALSE)

plot(hampelPsi)# calling the plot method (nicely showing "all" !)

[Package robustbase version 0.99-7 Index]