summary.lmrob {robustbase}	R Documentation

Summary Method for "lmrob" Objects

Description

Summary method for R object of class "lmrob" and print method for the summary object.

Further, methods fitted(), residuals() work (via the default methods), and predict() (see predict.lmrob, vcov(), weights() (see weights.lmrob), model.matrix(), confint(), dummy.coef(), hatvalues(), etc., have explicitly defined lmrob methods. .lmrob.hat() is the lower level “work horse” of the hatvalues() method.

Usage

## S3 method for class 'lmrob'
summary(object, correlation = FALSE,
        symbolic.cor = FALSE, ...)
## S3 method for class 'summary.lmrob'
print(x, digits = max(3, getOption("digits") - 3),
      symbolic.cor= x$symbolic.cor,
      signif.stars = getOption("show.signif.stars"),
      showAlgo = TRUE, ...)

## S3 method for class 'lmrob'
vcov(object, cov = object$control$cov, complete = TRUE, ...)
## S3 method for class 'lmrob'
model.matrix(object, ...)

Arguments

object
an R object of class lmrob, typically created by lmrob.

correlation
logical variable indicating whether to compute the correlation matrix of the estimated coefficients.

symbolic.cor
logical indicating whether to use symbols to display the above correlation matrix.

x
an R object of class summary.lmrob, typically resulting from summary(lmrob(..),..).

digits
number of digits for printing, see digits in options.

signif.stars
logical variable indicating whether to use stars to display different levels of significance in the individual t-tests.

showAlgo
optional logical indicating if the algorithmic parameters (as mostly inside the control part) should be shown.

cov
covariance estimation function to use, a function or character string naming the function; robustbase currently provides ".vcov.w" and ".vcov.avar1", see Details of lmrob. Particularly useful when object is the result of lmrob(.., cov = "none"), where

  object$cov <- vcov(object, cov = ".vcov.w")

allows to update the fitted object.

complete
(mainly for R >= 3.5.0:) logical indicating if the full variance-covariance matrix should be returned also in case of an over-determined system where some coefficients are undefined and coef(.) contains NAs correspondingly. When complete = TRUE, vcov() is compatible with coef() also in this singular case.

...
potentially more arguments passed to methods.

Value

summary(object) returns an object of S3 class "summary.lmrob", basically a list with components "call", "terms", "residuals", "scale", "rweights", "converged", "iter", "control" all copied from object, and further components, partly for compatibility with summary.lm,

coefficients
a matrix with columns "Estimate", "Std. Error", "t value", and "PR(>|t|)", where "Estimate" is identical to coef(object). Note that coef(<summary.obj>) is slightly preferred to access this matrix.

df
degrees of freedom, in an lm compatible way.

sigma
identical to sigma(object).

aliased
..

cov
derived from object$cov.

r.squared
robust “R squared” or R^2, a coefficient of determination: This is the consistency corrected robust coefficient of determination by Renaud and Victoria-Feser (2010).

adj.r.squared
an adjusted R squared, see r.squared.

References

Renaud, O. and Victoria-Feser, M.-P. (2010). A robust coefficient of determination for regression, Journal of Statistical Planning and Inference 140, 1852-1862.

See Also

lmrob, predict.lmrob, weights.lmrob, summary.lm, print, summary.

Examples

mod1 <- lmrob(stack.loss ~ ., data = stackloss)
sa <- summary(mod1)  # calls summary.lmrob(....)
sa                   # dispatches to call print.summary.lmrob(....)

## correlation between estimated coefficients:
cov2cor(vcov(mod1))

cbind(fit = fitted(mod1), resid = residuals(mod1),
      wgts= weights(mod1, type="robustness"),
      predict(mod1, interval="prediction"))

data(heart)
sm2 <- summary( m2 <- lmrob(clength ~ ., data = heart) )
sm2

[Package robustbase version 0.95-0 Index]