summary.lts {robustbase}	R Documentation

Summary Method for LTS objects

Description

summary method for class "lts".

Usage

## S3 method for class 'lts'
summary(object, correlation = FALSE, ...)

## S3 method for class 'summary.lts'
print(x, digits = max(3, getOption("digits") - 3),
     signif.stars = getOption("show.signif.stars"), ...)

Arguments

object
an object of class "lts", usually, a result of a call to ltsReg.

correlation
logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.

x
an object of class "summary.lts", usually, a result of a call to summary.lts.

digits
the number of significant digits to use when printing.

signif.stars
logical indicating if “significance stars” should be printer, see printCoefmat.

...
further arguments passed to or from other methods.

Details

These functions compute and print summary statistics for weighted least square estimates with weights based on LTS estimates. Therefore the statistics are similar to those for LS but all terms are multiplied by the corresponding weight.

Correlations are printed to two decimal places: to see the actual correlations print summary(object)$correlation directly.

Value

The function summary.lts computes and returns a list of summary statistics of the fitted linear model given in object, using the components of this object (list elements).

residuals
the residuals - a vector like the response y containing the residuals from the weighted least squares regression.

coefficients
a p x 4 matrix with columns for the estimated coefficient, its standard error, t-statistic and corresponding (two-sided) p-value.

sigma
the estimated scale of the reweighted residuals

sigma^2 = 1/(n-p) Sum(R[i]^2),

where R[i] is the i-th residual, residuals[i].

df
degrees of freedom, a 3-vector (p, n-p, p*), the last being the number of non-aliased coefficients.

fstatistic
(for models including non-intercept terms) a 3-vector with the value of the F-statistic with its numerator and denominator degrees of freedom.

r.squared
R^2, the “fraction of variance explained by the model”,

R^2 = 1 - Sum(R[i]^2) / Sum((y[i]- y*)^2),

where y* is the mean of y[i] if there is an intercept and zero otherwise.

adj.r.squared
the above R^2 statistic “adjusted”, penalizing for higher p.

cov.unscaled
a p x p matrix of (unscaled) covariances of the coef[j], j=1, …, p.

correlation
the correlation matrix corresponding to the above cov.unscaled, if correlation = TRUE is specified.

See Also

ltsReg; the generic summary.

Examples

data(Animals2)
ltsA <- ltsReg(log(brain) ~ log(body), data = Animals2)
(slts <- summary(ltsA))
## non-default options for printing the summary:
print(slts, digits = 5, signif.stars = FALSE)

[Package robustbase version 0.95-0 Index]