adjboxStats {robustbase}	R Documentation

Statistics for Skewness-adjusted Boxplots

Description

Computes the “statistics” for producing boxplots adjusted for skewed distributions as proposed in Hubert and Vandervieren (2004).

Usage

adjboxStats(x, coef = 1.5, a = -4, b = 3, do.conf = TRUE, do.out = TRUE)

Arguments

x 	 ~~Describe x here~~

coef 	 ~~Describe coef here~~

a 	 ~~Describe a here~~

b 	 ~~Describe b here~~

do.conf 	 ~~Describe do.conf here~~

do.out 	 ~~Describe do.out here~~

Value

A list with the components

stats 	 a matrix, each column contains the extreme of the lower whisker, the lower hinge, the median, the upper hinge and the extreme of the upper whisker for one group/plot. If all the inputs have the same class attribute, so will this component.

n 	 a vector with the number of observations in each group.

coef 	 a matrix where each column contains the lower and upper extremes of the notch.

out 	 the values of any data points which lie beyond the extremes of the whiskers.

Note

The code slightly modifies the code of boxplot.stats.

Author(s)

R Core Development Team, slightly adapted by Tobias Verbeke

See Also

adjbox(), also for references, the function which mainly uses this one; further boxplot.stats.

Examples

data(condroz)
astat <- adjboxStats(condroz[,"Ca"])
astat

[Package robustbase version 0.4-5 Index]