chgDefaults-methods {robustbase}	R Documentation

Methods for Function chgDefaults() in Package ‘robustbase’

Description

This is auxiliary functionality for modifying psi_func-class objects.

Methods

object = "psi_func"

The method is used to change the default values for the tuning parameters, and returns a psi_func-class object, a copy of input object with the slot tDefs possibly changed;.

Examples

## FIXME -- add one 

[Package robustbase version 0.4-5 Index]