estimethod {robustbase}	R Documentation

Extract the Estimation Method 'Estimethod' from a Fitted Model

Description

Extract the estimation method as a character string from a fitted model.

Usage

  estimethod(object, ...)

Arguments

object
a fitted model.

...
additional, optional arguments. (None are used in our methods)

Details

This is a (S3) generic function for which we provide methods, currently for nlrob only.

Value

a character string, the estimation method used.

See Also

nlrob, and nlrob.MM, notably for examples.

[Package robustbase version 0.95-0 Index]