glmrob..control {robustbase}	R Documentation

Controlling Robust GLM Fitting by Different Methods

Description

These are auxiliary functions as user interface for glmrob fitting when the different methods, "Mqle", "BY", or "MT" are used. Typically only used when calling glmrob.

Usage

glmrobMqle.control(acc = 1e-04, test.acc = "coef", maxit = 50, tcc = 1.345)
glmrobBY.control  (maxit = 1000, const = 0.5, maxhalf = 10)
glmrobMT.control  (cw = 2.1, nsubm = 500, acc = 1e-06, maxit = 200)

Arguments

acc
positive convergence tolerance; the iterations converge when ???

test.acc
Only "coef" is currently implemented

maxit
integer giving the maximum number of iterations.

tcc
tuning constant c for Huber's psi-function

const
for "BY", the normalizing constant ..

maxhalf
for "BY"; the number of halving steps when the gradient itself no longer improves. We have seen examples when increasing maxhalf was of relevance.

cw
tuning constant c for Tukey's biweight psi-function

nsubm
the number of subsamples to take for finding an initial estimate for method = "MT".

Value

A list with the arguments as components.

Author(s)

Andreas Ruckstuhl and Martin Maechler

See Also

glmrob

Examples

str(glmrobMqle.control())
str(glmrobBY.control())
str(glmrobMT.control())

[Package robustbase version 0.95-0 Index]