lmrob..D..fit {robustbase}	R Documentation

Compute Design Adaptive Scale estimate

Description

This function calculates a Design Adaptive Scale estimate for a given MM-estimate. This is supposed to be a part of a chain of estimates like SMD or SMDM.

Usage

lmrob..D..fit(obj, x=obj$x, control = obj$control,
              mf = obj$model, method = obj$control$method)

Arguments

obj
lmrob-object based on which the estimate is to be calculated.

x
the design matrix; if missing, the method tries to get it from obj$x and if this fails from obj$model.

control
list of control parameters, as returned by lmrob.control.

mf
unused and deprecated.

method
optional; the method used for obj computation.

Details

This function is used by lmrob.fit and typically not to be used on its own. Note that lmrob.fit() specifies control potentially differently than the default, but does use the default for method.

Value

The given lmrob-object with the following elements updated:

scale
The Design Adaptive Scale estimate

converged
TRUE if the scale calculation converged, FALSE other.

Author(s)

Manuel Koller

References

Koller, M. and Stahel, W.A. (2011), Sharpening Wald-type inference in robust regression for small samples, Computational Statistics & Data Analysis 55(8), 2504–2515.

See Also

lmrob.fit, lmrob

Examples

data(stackloss)
## Compute manual SMD-estimate:
## 1) MM-estimate
m1 <- lmrob(stack.loss ~ ., data = stackloss)
## 2) Add Design Adaptive Scale estimate
m2 <- lmrob..D..fit(m1)
print(c(m1$scale, m2$scale))

summary(m1)
summary(m2) ## the covariance matrix estimate is also updated

[Package robustbase version 0.95-0 Index]