lmrob.control {robustbase}	R Documentation

Tuning parameters for lmrob

Description

Tuning parameters for, lmrob, the MM-regression estimator and the associated S-estimator.

Usage

lmrob.control(seed = NULL, nResample = 500,
              tuning.chi = 1.54764, bb = 0.5, tuning.psi = 4.685061,
              max.it = 50, groups = 5, n.group = 400,
              k.fast.s = 1, best.r.s = 2, k.max = 200,
              refine.tol = 1e-7, rel.tol = 1e-7,
              trace.lev = 0, compute.rd = FALSE)

Arguments

seed 	 an integer vector, the seed to be used for random re-sampling used in obtaining candidates for the initial S-estimator; see .Random.seed. The current value of .Random.seed will be preserved if seed is set; otherwise (by default), .Random.seed will be modified as usual from calls to runif().

nResample 	 number of re-sampling candidates to be used to find the initial S-estimator. Currently defaults to 500 which works well in most situations (see references).

tuning.chi 	 tuning constant for the S-estimator. The default, 1.54764, yields a 50% breakdown estimator.

bb 	 expected value under the normal model of the “chi” (rather rho) function with tuning constant equal to tuning.chi. This is used to compute the S-estimator.

tuning.psi 	 tuning constant for the re-descending M-estimator. The choice 4.685061 yields an estimator with asymptotic efficiency of 95% for normal errors.

max.it 	 integer specifying the maximum number of IRWLS iterations.

groups 	 (for the fast-S algorithm): Number of random subsets to use when the data set is large.

n.group 	 (for the fast-S algorithm): Size of each of the groups above. Note that this must be at least p.

k.fast.s 	 (for the fast-S algorithm): Number of local improvement steps (“I-steps”) for each re-sampling candidate.

best.r.s 	 (for the fast-S algorithm): Number of of best candidates to be iterated further (i.e., “refined”); is denoted t in Salibian-Barrera & Yohai(2006).

k.max 	 (for the fast-S algorithm): maximal number of refinement steps for the “fully” iterated best candidates.

refine.tol 	 (for the fast-S algorithm): relative convergence tolerance for the fully iterated best candidates.

rel.tol 	 (for the RWLS iterations of the MM algorithm): relative convergence tolerance for the parameter vector.

trace.lev 	 integer indicating if the progress of the MM-algorithm should be traced (increasingly); default trace.lev = 0 does no tracing.

compute.rd 	 logical indicating if robust distances (based on the MCD robust covariance estimator covMcd) are to be computed for the robust diagnostic plots. This may take some time to finish, particularly for large data sets, and can lead to singularity problems when there are factor explanatory variables (with many levels, or levels with “few” observations). Hence, is FALSE by default.

Author(s)

Matias Salibian-Barrera and Martin Maechler

See Also

lmrob, also for references and examples.

Examples

## Show the default settings:
str(lmrob.control())

[Package robustbase version 0.4-5 Index]