plot.lts {robustbase}	R Documentation

Robust LTS Regression Diagnostic Plots

Description

Four plots (selectable by which) are currently provided:

a plot of the standardized residuals versus their index,
a plot of the standardized residuals versus fitted values,
a Normal Q-Q plot of the standardized residuals, and
a regression diagnostic plot (standardized residuals versus robust distances of the predictor variables).

Usage

## S3 method for class 'lts':
plot(x, which = c("all","rqq","rindex","rfit","rdiag"),
     classic=FALSE, ask=(which=="all" && dev.interactive()), id.n, ...)

Arguments

x 	 a lts object, typically result of ltsReg.

which 	 string indicating which plot to show. See the Details section for a description of the options. Defaults to "all".

classic 	 whether to plot the classical distances too. Default is FALSE.

ask 	 logical indicating if the user should be asked before each plot, see par(ask=.). Defaults to which == "all" && dev.interactive().

id.n 	 number of observations to be identified by a label starting with the most extreme. Default is the number of identified outliers (can be different for the different plots - see Details).

... 	 other parameters to be passed through to plotting functions.

Details

This function produces several plots based on the robust and classical regression estimates. Which of them to select is specified by the attribute which. The possible options are:

rqq:

Normal Q-Q plot of the standardized residuals;

rindex:

plot of the standardized residuals versus their index;

rfit:

plot of the standardized residuals versus fitted values;

rdiag:

regression diagnostic plot.

The normal quantile plot produces a normal Q-Q plot of the standardized residuals. A line is drawn which passes through the first and third quantile. The id.n residuals with largest distances from this line are identified by labels (the observation number). The default for id.n is the number of regression outliers (lts.wt==0).

In the Index plot and in the Fitted values plot the standardized residuals are displayed against the observation number or the fitted value respectively. A horizontal dashed line is drawn at 0 and two solid horizontal lines are located at +2.5 and -2.5. The id.n residuals with largest absolute values are identified by labels (the observation number). The default for id.n is the number regression outliers (lts.wt==0).

The regression diagnostic plot, introduced by Rousseeuw and van Zomeren (1990), displays the standardized residuals versus robust distances. Following Rousseeuw and van Zomeren (1990), the horizontal dashed lines are located at +2.5 and -2.5 and the vertical line is located at the upper 0.975 percent point of the chi-squared distribution with p degrees of freedom. The id.n residuals with largest absolute values and/or largest robust Mahalanobis distances are identified by labels (the observation number). The default for id.n is the number of all outliers: regression outliers (lts.wt==0) + leverage (bad and good) points (RD > 0.975 percent point of the chi-squared distribution with p degrees of freedom).

References

P. J. Rousseeuw and van Zomeren, B. C. (1990). Unmasking Multivariate Outliers and Leverage Points. Journal of the American Statistical Association 85, 633–639.

P. J. Rousseeuw and K. van Driessen (1999) A fast algorithm for the minimum covariance determinant estimator. Technometrics 41, 212–223.

See Also

covPlot

Examples

data(hbk)
lts <- ltsReg(Y ~ ., data = hbk)
lts
plot(lts, which = "rqq") 

[Package robustbase version 0.4-5 Index]