nlrob.control {robustbase}	R Documentation

Control Nonlinear Robust Regression Algorithms

Description

Allow the user to specify details for the different nonlinear robust regression algorithms in nlrob.

Usage

nlrob.control(method,
              psi = c("bisquare", "lqq", "welsh", "optimal", "hampel", "ggw"),
              init = c("S", "lts"),
              optimizer = "JDEoptim", optArgs  = list(),
              ...)

Arguments

method
character string specifying the method

psi
string specifying the psi-function which defines the estimator.

init
for some methods, currently, "MM" only, a string specifying the initial estimator.

optimizer
currently only "JDEoptim" from package DEoptimR.

optArgs
a list of optional arguments to the optimizer. Currently, that is JDEoptim from package DEoptimR.

...

Value

a list with several named components. The contents depend quite a bit on the method.

See Also

nlrob, nlrob,

Examples

str(nlrob.control("MM"))
str(nlrob.control("tau"))
str(nlrob.control("CM"))
str(nlrob.control("mtl"))

[Package robustbase version 0.95-0 Index]