psiFunc {robustbase}	R Documentation

Constructor for Objects "Psi Function" Class

Description

psiFunc(..) is a convenience interface to new("psi_func",..), i.e. for constructing objects of class "psi_func".

Usage

psiFunc(rho, psi, wgt, Dpsi, Erho = NULL, Epsi2 = NULL, EDpsi = NULL, ...)

Arguments

rho, psi, wgt, Dpsi 	 each a function of x and tuning parameters typically.

Erho, Epsi2, EDpsi 	 see psi_func-class, and note that these may change in the future.

... 	 potential further arguments for specifying tuning parameter names and defaults.

Author(s)

Martin Maechler

See Also

psi_func-class for the class description.

Examples

## classical {trivial, not interesting}:
F1 <- function(x) rep.int(1, length(x))
cPsi <- psiFunc(rho = function(x) x^2 / 2, psi = function(x) x,
                wgt = F1, Dpsi = F1,
                Erho = function(x) rep.int(1/2, length(x)),
                Epsi2 = F1, EDpsi = F1)

[Package robustbase version 0.4-5 Index]