psi_func-class {robustbase}	R Documentation

Class of "Psi Functions" for M-Estimation

Description

The class "psi_func" is used to store psi functions for M-estimation. In particular, an object of the class contains rho(x), its derivative psi(x), the weight function ψ(x)/x, and first derivative of ψ, Dpsi = ψ'(x).

Objects from the Class

Objects can be created by calls of the form new("psi_func", ...), but preferably by psiFunc(...).

Slots

rho:

the rho() function, an object of class "functionX". This is used to formulate the objective function; ρ() can be regarded as generalized negative log-likelihood.

psi:

psi() is the derivative of ρ, psi(x) = d/dx rho(x); also of class "functionX".

wgt:

The weight function ψ(x)/x, of class "functionX".

Dpsi:

the derivative of ψ, Dpsi(x) = psi'(x); of class "functionX".

Dwgt:

the derivative of the weight function, of class "functionX", is generated automatically if psiFunc constructor is used.

tDefs:

named numeric vector of tuning parameter Default values.

Erho:

A function of class "functionXal" for computing E[ρ(X)] when X is standard normal N(0,1).

Epsi2:

A function of class "functionXal" for computing E[ψ^2(X)] when X is standard normal.

EDpsi:

A function of class "functionXal" for computing E[ψ'(X)] when X is standard normal.

name:

Name of psi-function used for printing.

xtras:

Potentially further information.

Methods

Currently, only chgDefaults(), plot() and show().

Author(s)

Martin Maechler

See Also

psiFunc.

Examples

str(huberPsi, give.attr = FALSE)

plot(hampelPsi)# calling the plot method (nicely showing "all" !)

[Package robustbase version 0.95-0 Index]