psi_func-class {robustbase}	R Documentation

Class of "Psi Functions" for M-Estimation

Description

The class "psi_func" is used to store psi functions for M-estimation. In particular, an object of the class contains rho(x), its derivative psi(x), the weight function psi(x)/x, and first derivative of psi, Dpsi = psi'(x).

Objects from the Class

Objects can be created by calls of the form new("psi_func", ...), but preferably by psiFunc(...).

Slots

rho:

the rho() function, an object of class "functionX". This is used to formulate the objective function; rho() can be regarded as generalized negative log-likelihood.

psi:

psi() is the derivative of rho, psi(x) = d/dx rho(x); also of class "functionX".

wgt:

The weight function psi(x)/x, of class "functionX".

Dpsi:

the derivative of psi, Dpsi(x) = psi'(x); of class "functionX".

tDefs:

named numeric vector of tuning parameter Default values.

Erho:

Object of class "functionXal" ~~

Epsi2:

Object of class "functionXal" ~~

EDpsi:

Object of class "functionXal" ~~

xtras:

Potentially further information.

Methods

No methods defined with class "psi_func" in the signature.

Author(s)

Martin Maechler

See Also

psiFunc.

Examples

str(huberPsi, give.attr = FALSE)

[Package robustbase version 0.4-5 Index]