rrcov.control {robustbase}	R Documentation

Control object for the estimation parameters

Description

Auxilary function for passing the estimation options as parameters to the estimation functions.

NOTE: The name WILL change !!!!

Usage

rrcov.control(alpha = 1/2, nsamp = 500, seed = NULL, tolSolve = 1e-14,
              trace = FALSE,
              use.correction = TRUE, adjust = FALSE)

Arguments

alpha 	 This parameter controls the size of the subsets over which the determinant is minimized, i.e., alpha*n observations are used for computing the determinant. Allowed values are between 0.5 and 1 and the default is 0.5.

nsamp 	 number of subsets used for initial estimates or "best" or "exact". Default is nsamp = 500. If nsamp="best" exhaustive enumeration is done, as far as the number of trials do not exceed 5000. If nsamp="exact" exhaustive enumeration will be attempted however many samples are needed. In this case a warning message will be displayed saying that the computation can take a very long time.

seed 	 initial seed for R's random number generator; see .Random.seed and the description of the seed argument in lmrob.control.

tolSolve 	 numeric tolerance to be used for inversion (solve) of the covariance matrix in mahalanobis.

trace 	 whether to print intermediate results. Default is trace = FALSE

use.correction 	 whether to use finite sample correction factors. Defaults to TRUE.

adjust 	 whether to perform intercept adjustment at each step. Because this can be quite time consuming, the default is adjust = FALSE.

Details

For details about the estimation options see the corresponding estimation functions.

Value

A list with components, as the parameters passed by the invocation

Author(s)

Valentin Todorov

Examples

data(Animals, package = "MASS")
brain <- Animals[c(1:24, 26:25, 27:28),]
data(hbk)
hbk.x <- data.matrix(hbk[, 1:3])

ctrl <- rrcov.control(alpha=0.75, trace=TRUE)
covMcd(hbk.x,      control = ctrl)
covMcd(log(brain), control = ctrl)

[Package robustbase version 0.4-5 Index]