sigma {robustbase}	R Documentation

Extract 'Sigma' - Standard Deviation of Errors for Robust Models

Description

Extract the estimated standard deviation of the errors, the “residual standard deviation” (misnomed also “residual standard error”) from a fitted model.

Usage

## S3 method for class 'lmrob'
sigma(object, ...)

Arguments

object
a fitted model.

...
additional, optional arguments. (None are used in our methods)

Details

For R <= 3.2.x, we provide an (S3) generic function (as e.g., package lme4) and methods for lmrob, nlrob, and nls.

From R >= 3.3.0, we provide methods for our lmrob and nlrob models.

Value

the residual standard error as a scalar

Examples

      m.cl <-   lm (Y ~ ., data=coleman)
if(getRversion() >= "3.3.0")  sigma(m.cl)  else  summary(m.cl)$sigma
sigma( m1  <- lmrob(Y ~ ., data=coleman) )
sigma( m2  <- lmrob(Y ~ ., data=coleman, setting = "KS2014") )

[Package robustbase version 0.95-0 Index]