summarizeRobWeights {robustbase}	R Documentation

Print a Nice "summary" of Robustness Weights

Description

Print a nice “summary” about a numeric vector of robustness weights. Observations with weights around zero are marked as outliers.

Usage

summarizeRobWeights(w, digits = getOption("digits"),
                    header = "Robustness weights:",
                    eps = 0.1 / length(w), eps1 = 1e-3, ...)

Arguments

w 	 numeric vector of robustness weigths.

digits 	 digits to be used for printing.

header 	 string to be printed as header line.

eps 	 numeric tolerance eps: values of w with |w[i]| < eps/n are said to be outliers.

eps1 	 numeric tolerance: values of w with |1 - w[i]| < eps1 are said to have weight ‘~= 1’.

... 	 potential further arguments, passed to print().

Value

none; the function is used for its side effect of printing.

Author(s)

Martin Maechler

See Also

The summary methods for lmrob and glmrob make use of summarizeRobWeights().

Examples

w <- c(1,1,1,1,0,1,1,1,1,0,1,1,.9999,.99999, .5,.6,1e-12)
summarizeRobWeights(w) # two outside ~= {0,1}
summarizeRobWeights(w, eps1 = 5e-5)# now three outside {0,1}

## See the summary(<lmrob>) outputs

[Package robustbase version 0.4-5 Index]