summary.glmrob {robustbase}	R Documentation

Summarizing Robust Fits of Generalized Linear Models

Description

The summary method for class "glmrob" summarizes robust fits of (currently only discrete) generalized linear models.

Usage

## S3 method for class 'glmrob':
summary(object, correlation = FALSE, symbolic.cor = FALSE, ...)
## S3 method for class 'glmrob':
vcov(object, ...)

## S3 method for class 'summary.glmrob':
print(x, digits = max(3, getOption("digits") - 3),
      symbolic.cor = x$symbolic.cor,
      signif.stars = getOption("show.signif.stars"), ...)

Arguments

object 	 an object of class "glmrob", usually, a result of a call to glmrob.

correlation 	 logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.

symbolic.cor 	 logical. If TRUE, print the correlations in a symbolic form (see symnum) rather than as numbers.

... 	 further arguments passed to or from other methods.

x 	 an object of class "summary.glrob".

digits 	 the number of digits to use for printing.

signif.stars 	 logical indicating if the P-values should be visualized by so called “significance stars”.

Details

summary.glmrob returns an object of class "summary.glmrob".

Its print() method tries to be smart about formatting the coefficients, standard errors, etc, and gives “significance stars” if signif.stars is TRUE (as per default when options where not changed).

Value

The function summary.glmrob computes and returns a list of summary statistics of the robustly fitted linear model given in object. The following elements are in the list:

... 	 FIXME

Author(s)

Andreas Ruckstuhl

See Also

glmrob; the generic summary and also summary.glm.

Examples

data(epilepsy)
Rmod <- glmrob(Ysum ~ Age10 + Base4*Trt, family = poisson,
               data = epilepsy, method= "Mqle")
ss <- summary(Rmod)
ss ## calls print.summary.glmrob()
str(ss) ## internal STRucture of summary object

[Package robustbase version 0.4-5 Index]