summary.lmrob {robustbase}	R Documentation

Summary Method for "lmrob" Objects

Description

Summary method for R object of class "lmrob" and print method for the summary object.

Further, methods fitted(), residuals() or weights() work (via the default methods), and predict(), vcov(), model.matrix() have explicitly defined lmrob methods.

Usage

## S3 method for class 'lmrob':
summary(object, correlation = FALSE,
        symbolic.cor = FALSE, ...)
## S3 method for class 'summary.lmrob':
print(x, digits = max(3, getOption("digits") - 3),
      symbolic.cor= x$symbolic.cor,
      signif.stars = getOption("show.signif.stars"), ...)

## S3 method for class 'lmrob':
vcov(object, ...)
## S3 method for class 'lmrob':
model.matrix(object, ...)

## S3 method for class 'lmrob':
predict(object, newdata = NULL, scale = NULL, ...)

Arguments

object 	 an R object of class lmrob, typically created by lmrob.

correlation 	 logical variable indicating whether to compute the correlation matrix of the estimated coefficients.

symbolic.cor 	 logical indicating whether to use symbols to display the above correlation matrix.

x 	 an R object of class summary.lmrob, typically resulting from summary(lmrob(..),..).

digits 	 number of digits for printing, see digits in options.

signif.stars 	 logical variable indicating whether to use stars to display different levels of significance in the individual t-tests.

newdata 	 data.frame with “new” data values at which prediction should be done; see predict.

scale 	 scale (estimate) to be used; see predict.lm.

... 	 potentially more arguments passed to methods.

See Also

lmrob, summary.lm, print, summary.

Examples

mod1 <- lmrob(stack.loss ~ ., data = stackloss)
sa <- summary(mod1)  # calls summary.lmrob(....)
sa                   # dispatches to call print.summary.lmrob(....)

## correlation between estimated coefficients:
cov2cor(vcov(mod1))

cbind(fit = fitted(mod1), resid = residuals(mod1),
      wgts= weights(mod1),
      predict(mod1, interval="prediction"))

data(heart)
sm2 <- summary( m2 <- lmrob(clength ~ ., data = heart) )
sm2

[Package robustbase version 0.4-5 Index]