summary.mcd {robustbase}	R Documentation

Summary Method for MCD objects

Description

summary method for class "mcd".

Usage

## S3 method for class 'mcd'
summary(object, ...)
## S3 method for class 'summary.mcd'
print(x, digits = max(3, getOption("digits") - 3),
     print.gap = 2, ...)

Arguments

object,x
an object of class "mcd" (or "summary.mcd"); usually, a result of a call to covMcd.

digits
the number of significant digits to use when printing.

print.gap
number of horizontal spaces between numbers; see also print.default.

...
further arguments passed to or from other methods.

Details

summary.mcd(), the S3 method, simply returns an (S3) object of class "summary.mcd" for which there's a print method:

print.summary.mcd prints summary statistics for the weighted covariance matrix and location estimates with weights based on MCD estimates. While the function print.mcd prints only the robust estimates of the location and the covariance matrix, print.summary.mcd will print also the correlation matrix (if requested in the call to covMcd with cor=TRUE), the eigenvalues of the covariance or the correlation matrix and the robust (“Mahalanobis”) distances.

Value

summary.mcd returns an summary.mcd object, whereas the print methods returns its first argument via invisible, as all print methods do.

See Also

covMcd, summary

Examples

data(Animals, package = "MASS")
brain <- Animals[c(1:24, 26:25, 27:28),]
lbrain <- log(brain)
summary(cLB <- covMcd(lbrain))

[Package robustbase version 0.95-0 Index]