summary.nlrob {robustbase}	R Documentation

Summarizing Robust Fits of Nonlinear Regression Models

Description

summary method for objects of class "nlrob", i.e., nlrob() results. Currently it only works for nlrob(*, method="M").

Usage

## S3 method for class 'nlrob'
summary(object, correlation = FALSE, symbolic.cor = FALSE, ...)

Arguments

object
an object of class "nlrob", usually, a result of a call to nlrob.

correlation
logical variable indicating whether to compute the correlation matrix of the estimated coefficients.

symbolic.cor
logical indicating whether to use symbols to display the above correlation matrix.

...
further arguments passed to or from other methods.

Value

The function summary.nlrob computes and returns an object of class "summary.nlrob" of summary statistics of the robustly fitted linear model given in object. There is a print method, print.summary.lmrob(), which nicely formats the output.

The result keeps a large part of object's components such as residuals, cov or w, and additionally contains

coefficients
the matrix of coefficients, standard errors and p-values.

correlation
if the correlation argument was true, the correlation matrix of the parameters.

Author(s)

Andreas Ruckstuhl

See Also

nlrob(), also for examples.

[Package robustbase version 0.95-0 Index]