tolEllipsePlot {robustbase}	R Documentation

Tolerance Ellipse Plot

Description

Plots the 0.975 tolerance ellipse of the bivariate data set x. The ellipse is defined by those data points whose distance is equal to the squareroot of the 0.975 chisquare quantile with 2 degrees of freedom.

Usage

tolEllipsePlot(x, m.cov = covMcd(x), cutoff = NULL, id.n = NULL,
               classic = FALSE, tol = 1e-07,
               xlab = "", ylab = "",
               main = "Tolerance ellipse (97.5%)",
               txt.leg = c("robust", "classical"),
               col.leg = c("red", "blue"),
               lty.leg = c("solid","dashed"))

Arguments

x 	 a two dimensional matrix or data frame.

m.cov 	 an object similar to those of class "mcd"; however only its components center and cov will be used. If missing, the MCD will be computed (via covMcd()).

cutoff 	 numeric distance needed to flag data points outside the ellipse.

id.n 	 number of observations to be identified by a label. If not supplied, the number of observations with distance larger than cutoff is used.

classic 	 whether to plot the classical distances as well, FALSE by default.

tol 	 tolerance to be used for computing the inverse, see solve. Defaults to 1e-7.

xlab, ylab, main 	 passed to plot.default.

txt.leg, col.leg, lty.leg 	 character vectors of length 2 for the legend, only used if classic = TRUE.

Author(s)

Peter Filzmoser, Valentin Todorov and Martin Maechler

See Also

covPlot which calls tolEllipsePlot() when desired. ellipsoidhull and predict.ellipsoid from package cluster.

Examples

data(hbk)
hbk.x <- data.matrix(hbk[, 1:3])
mcd <- covMcd(hbk.x)       # compute mcd in advance
## must be a 2-dimensional data set: take the first two columns :
tolEllipsePlot(hbk.x[,1:2])

## an "impressive" example:
data(telef)
tolEllipsePlot(telef, classic=TRUE)

[Package robustbase version 0.4-5 Index]