weights.lmrob {robustbase}	R Documentation

Extract Robustness and Model Weights

Description

weights() extracts robustness weights or fitting (or prior) weights from a lmrob or glmrob object.

Usage

## S3 method for class 'lmrob'
weights(object, type = c("prior", "robustness"), ...)

Arguments

object
an object of class "lmrob" or "glmrob", typically the result of a call to lmrob, or glmrob, respectively.

type
the type of weights to be returned. Either "prior" (default), or "robustness".

...
not used currently.

Details

The “prior weights” correspond to the weights specified using the “weights” argument when calling lmrob. The “robustness weights” are the weights assigned by the M-estimator of regression, ψ(r_i/S) / (r_i/S). The robust coefficient estimate then numericarlly corresponds to a weighted least squares fit using the product of both types of weights as weights.

Value

Weights extracted from the object object.

Author(s)

Manuel Koller and Martin Maechler.

See Also

lmrob, glmrob and weights

[Package robustbase version 0.95-0 Index]